requirements.txt
google-adk[eval]==1.15.0
pytest
pytest-asyncio
fastapi
uvicorn
requests